Fix incorrectly escaped singlequotes
As described in #4, the line

     vim.command('call setqflist([{"bufnr": bufnr(""), "lnum": '+lnum+',
     "text": \''+qfitem+'\'}], "a")')

can create an incorrectly formatted vim message when `qfitem` has a
single singlequote in it. This replaces `qfitem` with
`qfitem.replace("'", "''")`.

Note that this works due to an oddity of vimscript. In vimscript,
singlequote delimited strings are "literal strings" (:h literal-string).
In these strings, the way to escape a singlequote is with another
singlequote.

This resolves #4.